requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
base58==2.1.1
bitarray==2.6.0
certifi==2022.6.15
charset-normalizer==2.1.1
cytoolz==0.12.0
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.1
gunicorn==20.1.0
hexbytes==0.3.0
idna==3.3
importlib-resources==5.9.0
ipfshttpclient==0.8.0a2
jsonschema==4.14.0
lru-dict==1.1.8
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
parsimonious==0.8.1
pkgutil_resolve_name==1.3.10
protobuf==3.20.1
pycryptodome==3.15.0
pyrsistent==0.18.1
pytz==2022.2.1
requests==2.28.1
rlp==2.0.1
six==1.16.0
sqlparse==0.4.2
toolz==0.12.0
urllib3==1.26.12
varint==1.0.2
web3==5.30.0
websockets==9.1
yarl==1.8.1
zipp==3.8.1
